requirements.txt
flask
mysql-connector-python
werkzeug
google-generativeai
